% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vg_lookup.R
\name{getENS_id}
\alias{getENS_id}
\title{getENS_id}
\usage{
getENS_id(name)
}
\arguments{
\item{name}{a vector containing the list of ENS Id's the user wishes to be cleaned}
}
\value{
a vector of searchable ENS Id's
}
\description{
This function helps clean up ENS Id's with unnecessary suffixes in them
}
