% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vg_lookup.R
\name{getVg}
\alias{getVg}
\title{getVg}
\usage{
getVg(Vg, colname = "IDs", tissue = "MSCLSK", ENS_id)
}
\arguments{
\item{Vg}{a data frame containing the population variance estimates (default is
VG_GTEx_v7.rda.}

\item{colname}{a string for the name of the column in the Vg file containing ENS
IDs (these will be used to look up the desired variance estimate).}

\item{tissue}{the column name in Vg from which the variance estimates should be
obtained. Here we default to MSCLSK.}

\item{ENS_id}{vector containing user's list of ENS IDs for which variance estimates
are desired.}
}
\value{
a dataframe containing the user inputted ENS_id column and the variance
estimates for those ENS IDs.
}
\description{
Function looks up the variance estimates from VG_GTEx_v7 data for a vector of ENS IDs
}
